chore: move path utils to shared for reuse
