+++ src/plugin/src/identifier-case/asset-rename-executor.js
@@ -11,7 +11,7 @@ import {
 
 import { isNonEmptyString } from "../../../shared/string-utils.js";
 import { isObjectLike } from "../../../shared/object-utils.js";
-import { fromPosixPath } from "../utils/path-utils.js";
+import { fromPosixPath } from "../../../shared/path-utils.js";
 import { DEFAULT_WRITE_ACCESS_MODE } from "./common.js";
 
 const defaultFsFacade = Object.freeze({
